Part 3. In this part, we have added a site map to the blog site.
Note:
Django comes with a sitemap framework, which allows you to
generate sitemaps for your site dynamically. A sitemap is an XML
file that tells search engines the pages of your website, their
relevance, and how frequently they are updated.

To implement this we edited the settings.py and added
i)django.contrib.sites and
ii)django.contrib.sitemaps.

After we included the above to site maps. Then made migrations.

2. To make this effective, we created a new file called sitemaps.py
3. We then linked it via the main url.

To see how the sitemaps dictionary is passed to the sitemap view
run run the development
server and open http://127.0.0.1:8000/sitemap.xml

4. To see the default object created, open
http://127.0.0.1:8000/admin/sites/site/ assuming 8000 is your port